Move `Tree` and `Frontier` traits to `incementalmerkletree::testing`
