+++ bridgetree/src/testing.rs
@@ -1,84 +1,11 @@
 mod complete_tree;
 
 use proptest::prelude::*;
-use std::collections::BTreeSet;
 
-use incrementalmerkletree::{testing::SipHashable, Hashable, Position};
-
-//
-// Traits used to permit comparison testing between tree implementations.
-//
-
-/// A possibly-empty incremental Merkle frontier.
-pub trait Frontier<H> {
-    /// Appends a new value to the frontier at the next available slot.
-    /// Returns true if successful and false if the frontier would exceed
-    /// the maximum allowed depth.
-    fn append(&mut self, value: &H) -> bool;
-
-    /// Obtains the current root of this Merkle frontier by hashing
-    /// against empty nodes up to the maximum height of the pruned
-    /// tree that the frontier represents.
-    fn root(&self) -> H;
-}
-
-/// A Merkle tree that supports incremental appends, marking of
-/// leaf nodes for construction of witnesses, checkpoints and rollbacks.
-pub trait Tree<H> {
-    /// Appends a new value to the tree at the next available slot.
-    /// Returns true if successful and false if the tree would exceed
-    /// the maximum allowed depth.
-    fn append(&mut self, value: &H) -> bool;
-
-    /// Returns the most recently appended leaf value.
-    fn current_position(&self) -> Option<Position>;
-
-    /// Returns the most recently appended leaf value.
-    fn current_leaf(&self) -> Option<&H>;
-
-    /// Returns the leaf at the specified position if the tree can produce
-    /// a witness for it.
-    fn get_marked_leaf(&self, position: Position) -> Option<&H>;
-
-    /// Marks the current leaf as one for which we're interested in producing
-    /// a witness. Returns an optional value containing the
-    /// current position if successful or if the current value was already
-    /// marked, or None if the tree is empty.
-    fn mark(&mut self) -> Option<Position>;
-
-    /// Return a set of all the positions for which we have marked.
-    fn marked_positions(&self) -> BTreeSet<Position>;
-
-    /// Obtains the root of the Merkle tree at the specified checkpoint depth
-    /// by hashing against empty nodes up to the maximum height of the tree.
-    /// Returns `None` if there are not enough checkpoints available to reach the
-    /// requested checkpoint depth.
-    fn root(&self, checkpoint_depth: usize) -> Option<H>;
-
-    /// Obtains a witness to the value at the specified position,
-    /// as of the tree state corresponding to the given root.
-    /// Returns `None` if there is no available witness to that
-    /// position or if the root does not correspond to a checkpointed
-    /// root of the tree.
-    fn witness(&self, position: Position, as_of_root: &H) -> Option<Vec<H>>;
-
-    /// Marks the value at the specified position as a value we're no longer
-    /// interested in maintaining a mark for. Returns true if successful and
-    /// false if we were already not maintaining a mark at this position.
-    fn remove_mark(&mut self, position: Position) -> bool;
-
-    /// Creates a new checkpoint for the current tree state. It is valid to
-    /// have multiple checkpoints for the same tree state, and each `rewind`
-    /// call will remove a single checkpoint.
-    fn checkpoint(&mut self);
-
-    /// Rewinds the tree state to the previous checkpoint, and then removes
-    /// that checkpoint record. If there are multiple checkpoints at a given
-    /// tree state, the tree state will not be altered until all checkpoints
-    /// at that tree state have been removed using `rewind`. This function
-    /// return false and leave the tree unmodified if no checkpoints exist.
-    fn rewind(&mut self) -> bool;
-}
+use incrementalmerkletree::{
+    testing::{SipHashable, Tree},
+    Hashable, Position,
+};
 
 //
 // Types and utilities for shared example tests.
